requirements.txt
# Web framework
streamlit>=1.40.0

# LLM – Google Gemini (API only, no web-search API used)
google-generativeai>=0.7.0

# HTTP client – modern, supports redirects and timeouts cleanly
httpx>=0.27.0

# HTML parsing and link extraction
beautifulsoup4>=4.12.3
lxml>=5.2.0

# Main-content extraction from web pages (no external API)
trafilatura>=1.9.0

# Environment variable loading
python-dotenv>=1.0.0
